Add an --only-if-exact command line option
This behaves like --skip-ui-if-exact, but skips the UI in all cases.
Because Vim needs to launch QuickOpen differently for curses, launching
the UI from within QuickOpen is too late to switch UI modes. So, instead,
bail out if no match was found.

I don't have enough knowledge of how emacs works to adjust its use of the
--skip-ui-if-exact flag, so I decided to introduce this option in parallel
in the short term.