Increase indexing speed by 70%
This patch replaces the previous indexing system (when possible) with a combination
of /usr/bin/find and a better ignore-rule processing.

On a macbook air, indexing of a 266k file chrome checkout goes from 111s to 33s.

Althought the gain of using find is substantial, it does return many more files
than we need. For example, in the old indexer, an ignore rule like .* would
cause us to skip the .git directory. In the new indexer, we get all those
files. So, the ultimate improvement of using find with the dir_cache's ignore
system is only 25s.

To address this, we manually handle the ignore rules. The key detail in the
ignore rules is that fnmatch is horribly slow, so memoziation techniques of
various kinds are highly effective. This gives us the balance of the
improvements.

Much more performance headroom remains. Disabling all ignore processing and
only doing  the find takes only 18s. A better way of handling the ignore rules
needs to be devised.