Add --base-path argument
The --base-path argument allows filtering the results based on the file path.
This is a convenient way to work with multiple repositories, indexing them all
but showing results only on the one being worked on.

To use it, just add --base-path=/path/to/src/repo to the `search` command. This
will filter the results list to show only those files within that path. Omitting
the argument disables filtering.

This patch also adds a default base-path to the quickopen.vim plugin. It'll look
for the root-most directory named "src" in vim's current working directory path
and use that as the base-path. If none is found, the argument isn't added.